@@ -8,7 +8,7 @@
     "build": "tsc",
     "watch": "tsc -w"
   },
-  "devDependencies": {
-    "typescript": "^3.3.3"
+  "dependencies": {
+    "typescript": "^4.8.4"
   }
 }
