@@ -2,10 +2,12 @@ var __extends = (this && this.__extends) || (function () {
     var extendStatics = function (d, b) {
         extendStatics = Object.setPrototypeOf ||
             ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
-            function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
+            function (d, b) { for (var p in b) if (Object.prototype.hasOwnProperty.call(b, p)) d[p] = b[p]; };
         return extendStatics(d, b);
     };
     return function (d, b) {
+        if (typeof b !== "function" && b !== null)
+            throw new TypeError("Class extends value " + String(b) + " is not a constructor or null");
         extendStatics(d, b);
         function __() { this.constructor = d; }
         d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
@@ -34,18 +36,18 @@ define(["require", "exports", "esri/core/accessorSupport/decorators", "esri/widg
         }
         Header.prototype.render = function () {
             var classes = {};
-            return (widget_1.tsx("div", { bind: this, class: CSS.base, classes: classes },
-                widget_1.tsx("div", { class: CSS.label }, this.title),
-                widget_1.tsx("div", { class: CSS.actionContent }, this.actionContent && this.actionContent.map(function (content) { return content.render(); }))));
+            return ((0, widget_1.tsx)("div", { bind: this, class: CSS.base, classes: classes },
+                (0, widget_1.tsx)("div", { class: CSS.label }, this.title),
+                (0, widget_1.tsx)("div", { class: CSS.actionContent }, this.actionContent && this.actionContent.map(function (content) { return content.render(); }))));
         };
         __decorate([
-            decorators_1.property()
+            (0, decorators_1.property)()
         ], Header.prototype, "title", void 0);
         __decorate([
-            decorators_1.property()
+            (0, decorators_1.property)()
         ], Header.prototype, "actionContent", void 0);
         Header = __decorate([
-            decorators_1.subclass("widgets.Header")
+            (0, decorators_1.subclass)("widgets.Header")
         ], Header);
         return Header;
     }(Widget));
