@@ -2,10 +2,12 @@ var __extends = (this && this.__extends) || (function () {
     var extendStatics = function (d, b) {
         extendStatics = Object.setPrototypeOf ||
             ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
-            function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
+            function (d, b) { for (var p in b) if (Object.prototype.hasOwnProperty.call(b, p)) d[p] = b[p]; };
         return extendStatics(d, b);
     };
     return function (d, b) {
+        if (typeof b !== "function" && b !== null)
+            throw new TypeError("Class extends value " + String(b) + " is not a constructor or null");
         extendStatics(d, b);
         function __() { this.constructor = d; }
         d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
@@ -17,7 +19,7 @@ var __decorate = (this && this.__decorate) || function (decorators, target, key,
     else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
     return c > 3 && r && Object.defineProperty(target, key, r), r;
 };
-define(["require", "exports", "esri/core/accessorSupport/decorators", "esri/widgets/Widget", "esri/widgets/support/widget", "esri/core/watchUtils"], function (require, exports, decorators_1, Widget, widget_1, watchUtils_1) {
+define(["require", "exports", "esri/core/accessorSupport/decorators", "esri/widgets/Widget", "esri/widgets/support/widget", "esri/core/reactiveUtils"], function (require, exports, decorators_1, Widget, widget_1, reactiveUtils_1) {
     "use strict";
     Object.defineProperty(exports, "__esModule", { value: true });
     var CSS = {
@@ -30,19 +32,19 @@ define(["require", "exports", "esri/core/accessorSupport/decorators", "esri/widg
     var Indicator = /** @class */ (function (_super) {
         __extends(Indicator, _super);
         function Indicator(props) {
-            var _this = _super.call(this) || this;
+            var _this = _super.call(this, props) || this;
             // promise to the current stats.
             _this._statsPromise = null;
             _this._refresh = false;
             _this.iconClass = "esri-icon-dashboard";
             return _this;
         }
-        Indicator.prototype.postInitialize = function () {
+        Indicator.prototype.initialize = function () {
             var _this = this;
             var updateCallback = function () { return _this.updateStatistics(); };
             this.own([
-                watchUtils_1.whenFalse(this, "view.updating", updateCallback),
-                watchUtils_1.watch(this, ["view.extent", "statisticDefinition", "geometry"], updateCallback)
+                (0, reactiveUtils_1.when)(function () { var _a; return !((_a = _this.view) === null || _a === void 0 ? void 0 : _a.updating); }, updateCallback),
+                (0, reactiveUtils_1.watch)(function () { var _a; return [(_a = _this.view) === null || _a === void 0 ? void 0 : _a.extent, _this.geometry]; }, updateCallback)
             ]);
         };
         Indicator.prototype.updateStatistics = function () {
@@ -96,35 +98,35 @@ define(["require", "exports", "esri/core/accessorSupport/decorators", "esri/widg
         };
         Indicator.prototype.render = function () {
             var classes = {};
-            return (widget_1.tsx("div", { bind: this, class: CSS.base, classes: classes },
-                widget_1.tsx("div", { class: CSS.titleWrapper },
-                    widget_1.tsx("span", { class: CSS.title }, this.title)),
-                widget_1.tsx("div", { class: CSS.valueWrapper },
-                    widget_1.tsx("span", { class: CSS.value }, !this.value ? "-" : this.format.format(this.value)))));
+            return ((0, widget_1.tsx)("div", { bind: this, class: CSS.base, classes: classes },
+                (0, widget_1.tsx)("div", { class: CSS.titleWrapper },
+                    (0, widget_1.tsx)("span", { class: CSS.title }, this.title)),
+                (0, widget_1.tsx)("div", { class: CSS.valueWrapper },
+                    (0, widget_1.tsx)("span", { class: CSS.value }, !this.value ? "-" : this.format.format(this.value)))));
         };
         __decorate([
-            decorators_1.property()
+            (0, decorators_1.property)()
         ], Indicator.prototype, "iconClass", void 0);
         __decorate([
-            decorators_1.property()
+            (0, decorators_1.property)()
         ], Indicator.prototype, "queryStatistics", void 0);
         __decorate([
-            decorators_1.property()
+            (0, decorators_1.property)()
         ], Indicator.prototype, "layer", void 0);
         __decorate([
-            decorators_1.property()
+            (0, decorators_1.property)()
         ], Indicator.prototype, "title", void 0);
         __decorate([
-            decorators_1.property()
+            (0, decorators_1.property)()
         ], Indicator.prototype, "value", void 0);
         __decorate([
-            decorators_1.property()
+            (0, decorators_1.property)()
         ], Indicator.prototype, "view", void 0);
         __decorate([
-            decorators_1.property()
+            (0, decorators_1.property)()
         ], Indicator.prototype, "geometry", void 0);
         Indicator = __decorate([
-            decorators_1.subclass("widgets.Indicator")
+            (0, decorators_1.subclass)("widgets.Indicator")
         ], Indicator);
         return Indicator;
     }(Widget));
