@@ -2,10 +2,12 @@ var __extends = (this && this.__extends) || (function () {
     var extendStatics = function (d, b) {
         extendStatics = Object.setPrototypeOf ||
             ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
-            function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
+            function (d, b) { for (var p in b) if (Object.prototype.hasOwnProperty.call(b, p)) d[p] = b[p]; };
         return extendStatics(d, b);
     };
     return function (d, b) {
+        if (typeof b !== "function" && b !== null)
+            throw new TypeError("Class extends value " + String(b) + " is not a constructor or null");
         extendStatics(d, b);
         function __() { this.constructor = d; }
         d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
@@ -41,14 +43,14 @@ define(["require", "exports", "esri/core/accessorSupport/decorators", "esri/widg
         //
         //--------------------------------------------------------------------------
         Slider.prototype.render = function () {
-            return (widget_1.tsx("div", { bind: this, class: CSS.base },
-                widget_1.tsx("span", { class: CSS.title }, this.title),
-                widget_1.tsx("div", { class: "aria-widget-slider" },
-                    widget_1.tsx("div", { class: "rail-label min" }, "0"),
-                    widget_1.tsx("div", { class: "rail", style: "width: 300px;" },
-                        widget_1.tsx("img", { id: "minPriceHotel", src: "images/min-arrow.png", role: "slider", tabindex: "0", class: "min thumb", "aria-valuemin": "0", "aria-valuenow": "100", "aria-valuetext": "$100", "aria-valuemax": "400", "aria-label": "Hotel Minimum Price" }),
-                        widget_1.tsx("img", { id: "maxPriceHotel", src: "images/max-arrow.png", role: "slider", tabindex: "0", class: "max thumb", "aria-valuemin": "0", "aria-valuenow": "250", "aria-valuetext": "$250", "aria-valuemax": "400", "aria-label": "Hotel Maximum Price" })),
-                    widget_1.tsx("div", { class: "rail-label max" }, "0"))));
+            return ((0, widget_1.tsx)("div", { bind: this, class: CSS.base },
+                (0, widget_1.tsx)("span", { class: CSS.title }, this.title),
+                (0, widget_1.tsx)("div", { class: "aria-widget-slider" },
+                    (0, widget_1.tsx)("div", { class: "rail-label min" }, "0"),
+                    (0, widget_1.tsx)("div", { class: "rail", style: "width: 300px;" },
+                        (0, widget_1.tsx)("img", { id: "minPriceHotel", src: "images/min-arrow.png", role: "slider", tabindex: "0", class: "min thumb", "aria-valuemin": "0", "aria-valuenow": "100", "aria-valuetext": "$100", "aria-valuemax": "400", "aria-label": "Hotel Minimum Price" }),
+                        (0, widget_1.tsx)("img", { id: "maxPriceHotel", src: "images/max-arrow.png", role: "slider", tabindex: "0", class: "max thumb", "aria-valuemin": "0", "aria-valuenow": "250", "aria-valuetext": "$250", "aria-valuemax": "400", "aria-label": "Hotel Maximum Price" })),
+                    (0, widget_1.tsx)("div", { class: "rail-label max" }, "0"))));
         };
         Slider.prototype._changeHandler = function (event) {
             this.value = parseFloat(event.target.value);
@@ -57,28 +59,28 @@ define(["require", "exports", "esri/core/accessorSupport/decorators", "esri/widg
             }
         };
         __decorate([
-            decorators_1.property()
+            (0, decorators_1.property)()
         ], Slider.prototype, "action", void 0);
         __decorate([
-            decorators_1.property()
+            (0, decorators_1.property)()
         ], Slider.prototype, "min", void 0);
         __decorate([
-            decorators_1.property()
+            (0, decorators_1.property)()
         ], Slider.prototype, "max", void 0);
         __decorate([
-            decorators_1.property()
+            (0, decorators_1.property)()
         ], Slider.prototype, "step", void 0);
         __decorate([
-            decorators_1.property()
+            (0, decorators_1.property)()
         ], Slider.prototype, "value", void 0);
         __decorate([
-            decorators_1.property()
+            (0, decorators_1.property)()
         ], Slider.prototype, "title", void 0);
         __decorate([
-            widget_1.accessibleHandler()
+            (0, widget_1.accessibleHandler)()
         ], Slider.prototype, "_changeHandler", null);
         Slider = __decorate([
-            decorators_1.subclass("widgets.Slider")
+            (0, decorators_1.subclass)("widgets.Slider")
         ], Slider);
         return Slider;
     }(Widget));
